ref: Move JingleApi to the jingle package.
