+++ jicofo/src/test/java/mock/util/UtilityJingleOpSet.java
@@ -23,8 +23,6 @@ import org.jitsi.jicofo.xmpp.jingle.*;
 import org.jitsi.utils.logging2.*;
 import org.jitsi.xmpp.extensions.jingle.*;
 
-import org.jitsi.protocol.xmpp.*;
-
 import org.jivesoftware.smack.*;
 import org.jivesoftware.smack.packet.*;
